Extract Multipart service methods to MultipartService
