+++ server/src/test/java/com/adobe/testing/s3mock/BucketControllerTest.java
@@ -39,11 +39,11 @@ import com.adobe.testing.s3mock.dto.Owner;
 import com.adobe.testing.s3mock.dto.S3Object;
 import com.adobe.testing.s3mock.dto.StorageClass;
 import com.adobe.testing.s3mock.service.BucketService;
+import com.adobe.testing.s3mock.service.MultipartService;
 import com.adobe.testing.s3mock.service.ObjectService;
 import com.adobe.testing.s3mock.store.BucketStore;
 import com.adobe.testing.s3mock.store.FileStore;
 import com.adobe.testing.s3mock.store.KmsKeyStore;
-import com.adobe.testing.s3mock.store.S3ObjectMetadata;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.dataformat.xml.XmlMapper;
 import java.nio.file.Paths;
@@ -63,7 +63,8 @@ import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
 
 @AutoConfigureWebMvc
 @AutoConfigureMockMvc
-@MockBean(classes = {KmsKeyStore.class, BucketStore.class, FileStore.class, ObjectService.class})
+@MockBean(classes = {KmsKeyStore.class, BucketStore.class, FileStore.class, ObjectService.class,
+    MultipartService.class})
 @SpringBootTest(classes = {S3MockConfiguration.class})
 class BucketControllerTest {
