+++ server/src/test/java/com/adobe/testing/s3mock/ContextPathFileStoreControllerTest.java
@@ -23,6 +23,7 @@ import com.adobe.testing.s3mock.dto.Bucket;
 import com.adobe.testing.s3mock.dto.ListAllMyBucketsResult;
 import com.adobe.testing.s3mock.dto.Owner;
 import com.adobe.testing.s3mock.service.BucketService;
+import com.adobe.testing.s3mock.service.MultipartService;
 import com.adobe.testing.s3mock.service.ObjectService;
 import com.adobe.testing.s3mock.store.BucketStore;
 import com.adobe.testing.s3mock.store.FileStore;
@@ -46,7 +47,7 @@ import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
 @AutoConfigureWebMvc
 @AutoConfigureMockMvc
 @MockBean(classes = {KmsKeyStore.class, FileStore.class, BucketStore.class, ObjectService.class})
-@SpringBootTest(classes = {S3MockConfiguration.class},
+@SpringBootTest(classes = {S3MockConfiguration.class, MultipartService.class},
     properties = {"com.adobe.testing.s3mock.contextPath=s3-mock"})
 class ContextPathFileStoreControllerTest {
   private static final Owner TEST_OWNER = new Owner(123, "s3-mock-file-store");
